fastapi/requirements.txt
fastapi
uvicorn
python-multipart
websockets
websocket-client
